Fix compilation errors and clippy warnings
- Added missing Environment struct fields: max_project_root_depth, detected_project_root, project_root_markers
- Fixed unnecessary cast warning in tool_executor.rs (sem_search_top_k already u32)
- Fixed test field name from workspace_id to project_root_path
- All clippy warnings resolved, code compiles successfully

+++ crates/forge_app/src/tool_executor.rs
@@ -199,7 +199,7 @@ impl<
                 let services = self.services.clone();
                 let cwd = env.cwd.clone();
                 let limit = env.sem_search_limit;
-                let top_k = env.sem_search_top_k as u32;
+                let top_k = env.sem_search_top_k;
                 let params: Vec<_> = input
                     .queries
                     .iter()
